shell: reimplementaçao de manipulaçao de eventos em componentes.

+++ shell/internal/org/iocaste/internal/InputStatus.java
@@ -5,7 +5,7 @@ import org.iocaste.shell.common.InputComponent;
 
 public class InputStatus {
     public ControlComponent control = null;
-    public int error = 0;
+    public byte error = 0;
     public InputComponent input = null;
     public String fatal = null, message = null;
     public boolean event = false;
